PRODDEV-509 sync rachet
